Move JsonableTree into the tree module
